requirements.txt
requests==2.28.1
pyTelegramBotAPI==4.7.0
python-telegram-bot-calendar==1.0.5
python-dotenv==0.21.0
peewee==3.15.3


